[CodeQL] MsGraphicsPkg CodeQL Fixes (#179)
# Preface

Please ensure you have read the [contribution
docs](https://github.com/microsoft/mu/blob/master/CONTRIBUTING.md) prior
to submitting the pull request. In particular,
[pull request
guidelines](https://github.com/microsoft/mu/blob/master/CONTRIBUTING.md#pull-request-best-practices).

## Description

CodeQL flagged several issues in this pkg. This PR fixes them up.

For each item, place an "x" in between `[` and `]` if true. Example:
`[x]`.
_(you can also check items in the GitHub UI)_

- [x] Impacts functionality?
- **Functionality** - Does the change ultimately impact how firmware
functions?
- Examples: Add a new library, publish a new PPI, update an algorithm,
...
- [x] Impacts security?
- **Security** - Does the change have a direct security impact on an
application,
    flow, or firmware?
  - Examples: Crypto algorithm change, buffer overflow fix, parameter
    validation improvement, ...
- [ ] Breaking change?
- **Breaking change** - Will anyone consuming this change experience a
break
    in build or boot behavior?
- Examples: Add a new library class, move a module to a different repo,
call
    a function in a new library class in a pre-existing module, ...
- [ ] Includes tests?
  - **Tests** - Does the change include any explicit test code?
  - Examples: Unit tests, integration tests, robot tests, ...
- [ ] Includes documentation?
- **Documentation** - Does the change contain explicit documentation
additions
    outside direct code modifications (and comments)?
- Examples: Update readme file, add feature readme file, link to
documentation
    on an a separate Web page, ...

## How This Was Tested

Tested in a local CI build.

## Integration Instructions

N/A.

+++ MsGraphicsPkg/DisplayEngineDxe/FormDisplay.c
@@ -2488,7 +2488,7 @@ UiDisplayMenu (
               // move old value to return value honoring the values of
               // Src and Tgt (ie MOVE)
               Jndex = 0;
-              for (Index = 0; Index < (Statement->CurrentValue.BufferLen / sizeof (UINT32)); ) {
+              for (Index = 0; Index < (UINT32)(Statement->CurrentValue.BufferLen / sizeof (UINT32)); ) {
                 if (Index == Tgt) {
                   ReturnValue[Index++] = ValueArray[Src];
                 } else if (Jndex == Src) {
